refactor: convert taskfile.Merge into a method
